Install AdminLTE via Bower
AdminLTE is a great dashboard template but it is built with pure HTML
and CSS. In this tutorial, I'd like to go through installing AdminLTE
with Laravel 5 like pros. We will make sure the we have the ability to
update AdminLTE to the latest version at any given time by using Bower
to manage it.